requirements.txt
MetaTrader5
pandas
requests
pytz
pyyaml